Update AMD RNN benchmark to use MIOpenRNNForwardInference
The original AMD RNN benchmark used MIOpenRNNForwardTraining for
the inference pass.  However, this is unnecessary because the
DeepBench inference-only pass can use MIOpenRNNForwardInference
instead and get the same answer.  Per the MIOpen documentation,
MIOpenRNNForwardTraining should be used when the intermediate data
required for training is needed.

With this change, the RNN inference-only results should see a
slight improvement.