Update Nvidia RNN benchmark to use cudnnRNNForwardInference
The original NVIDIA RNN benchmark used cudnnRNNForwardTraining for
the inference pass.  However, this is unnecessary because the
DeepBench inference-only pass can use cudnnRNNForwardInference
instead and get the same answer.  Per the cuDNN documentation,
cudnnRNNForwardTraining should be used when the intermediate data
required for training is needed.

With this change, the RNN inference-only results should see a
slight improvement.